+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [vNext]
+### Added
+- DDC staking initial impl in `ddc-staking` pallet
+
+### Changed
+- Node runtime extended with `ddc-staking` pallet
+
 ## [2.33.0]
 ### Changed
 - Updated Substrate to polkadot-v0.9.19